verhindere Löschen von Abteilungen in denen noch Mitarbeiter sind

+++ classes/DepartmentDb.php
@@ -8,12 +8,12 @@ class DepartmentDb extends Department
    */
   public function getObjectById(int $id): DepartmentDb
   {
-      $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
-      $sql = "SELECT * FROM department WHERE id = :id";
-      $stmt = $dbh->prepare($sql);
-      $stmt->bindParam('id', $id, PDO::PARAM_INT);
-      $stmt->execute();
-      return $stmt->fetchObject(__CLASS__);
+    $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
+    $sql = "SELECT * FROM department WHERE id = :id";
+    $stmt = $dbh->prepare($sql);
+    $stmt->bindParam('id', $id, PDO::PARAM_INT);
+    $stmt->execute();
+    return $stmt->fetchObject(__CLASS__);
   }
 
   /**
@@ -21,18 +21,18 @@ class DepartmentDb extends Department
    */
   public function getAllAsObjects(): array|null
   {
-      try {
-        $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
-        $sql = "SELECT * FROM department";
-        $stmt = $dbh->query($sql);
-        $departments = [];
-        while ($dep = $stmt->fetchObject(__CLASS__)) {
-          $departments[] = $dep;
-        }
-        $dbh = null;
-      } catch (PDOException $e) {
-        throw new PDOException('Datenbank sagt nein: ' . $e->getMessage());
+    try {
+      $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
+      $sql = "SELECT * FROM department";
+      $stmt = $dbh->query($sql);
+      $departments = [];
+      while ($dep = $stmt->fetchObject(__CLASS__)) {
+        $departments[] = $dep;
       }
+      $dbh = null;
+    } catch (PDOException $e) {
+      throw new PDOException('Datenbank sagt nein: ' . $e->getMessage());
+    }
     return $departments;
   }
 
@@ -42,26 +42,52 @@ class DepartmentDb extends Department
    */
   public function createNewObject(string $name): DepartmentDb
   {
-      $dbh = new PDO (DB_DSN, DB_USER, DB_PASSWD);
-      $sql = "INSERT INTO department (id, name) VALUES (NULL, :name)";
-      $stmt = $dbh->prepare($sql);
-      $stmt->bindParam('name', $name);
-      $stmt->execute();
-      $id = $dbh->lastInsertId();
-      return new DepartmentDb($id, $name);
+    $dbh = new PDO (DB_DSN, DB_USER, DB_PASSWD);
+    $sql = "INSERT INTO department (id, name) VALUES (NULL, :name)";
+    $stmt = $dbh->prepare($sql);
+    $stmt->bindParam('name', $name);
+    $stmt->execute();
+    $id = $dbh->lastInsertId();
+    return new DepartmentDb($id, $name);
   }
 
   /**
    * @param int $id
    * @return void
+   * @throws Exception
+   */
+//  public function delete(int $id): void
+//  {
+//    try {
+//      $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
+//      $sql = "DELETE FROM department WHERE id = :id";
+//      $stmt = $dbh->prepare($sql);
+//      $stmt->bindParam('id', $id, PDO::PARAM_INT);
+//      $stmt->execute();
+//    } catch (PDOException $e) {
+//      throw new Exception('Fehler in delete Department: ' . $e->getMessage());
+//    }
+//  }
+
+  /**
+   * @param int $id
+   * @return void
+   * @throws Exception
    */
   public function delete(int $id): void
   {
-      $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
-      $sql = "DELETE FROM department WHERE id = :id";
-      $stmt = $dbh->prepare($sql);
-      $stmt->bindParam('id', $id, PDO::PARAM_INT);
-      $stmt->execute();
+    $employeesLeft = (new EmployeeDb())->getAllEmployeesByDepartment((new DepartmentDb())->getObjectById($id));
+    if (count($employeesLeft) === 0) {
+      try {
+        $dbh = new PDO (DB_DSN, DB_USER, DB_PASSWD);
+        $sql = "DELETE FROM department WHERE id = :id";
+        $stmt = $dbh->prepare($sql);
+        $stmt->bindParam('id', $id, PDO::PARAM_INT);
+        $stmt->execute();
+      } catch (PDOException $e) {
+        throw new Exception('Fehler in delete Department: ' . $e->getMessage());
+      }
+    }
   }
 
   /**
@@ -70,17 +96,17 @@ class DepartmentDb extends Department
    */
   public function updateObject(): void
   {
-      try {
-        $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
-        $sql = "UPDATE department SET name = :name WHERE id = :id";
-        $stmt = $dbh->prepare($sql);
-        $stmt->bindParam('name', $this->name);
-        $stmt->bindParam('id', $this->id);
-        $stmt->execute();
-        $dbh = null;
-      } catch (PDOException $e) {
-        throw new Exception($e->getMessage());
-      }
+    try {
+      $dbh = new PDO(DB_DSN, DB_USER, DB_PASSWD);
+      $sql = "UPDATE department SET name = :name WHERE id = :id";
+      $stmt = $dbh->prepare($sql);
+      $stmt->bindParam('name', $this->name);
+      $stmt->bindParam('id', $this->id);
+      $stmt->execute();
+      $dbh = null;
+    } catch (PDOException $e) {
+      throw new Exception($e->getMessage());
+    }
   }
 
 }
